+++ server/comm/xen_imaging.php
@@ -86,7 +86,16 @@
 
 					exec( 'partprobe' );
 
-					$command = 'ssh root@' . $settings['xen_migrate_ip'] . $extrasshsettings . ' ' . '\'' . '(pv -n /dev/' . $dev . ' | dd of="' . $settings['mount_migrate'] . '/xen' . $disknumber . '.img" bs=1M conv=notrunc,noerror status=none) > ' . $progressfilename . ' 2>&1 &' . '\'';//trailing & sends to background
+
+					if (empty($settings['compress'])) {
+
+						$command = 'ssh root@' . $settings['xen_migrate_ip'] . $extrasshsettings . ' ' . '\'' . '(pv -n /dev/' . $dev . ' | dd of="' . $settings['mount_migrate'] . '/xen' . $disknumber . '.img" bs=1M conv=notrunc,noerror status=none) > ' . $progressfilename . ' 2>&1 &' . '\'';//trailing & sends to background
+
+					} else {
+
+						$command = 'ssh root@' . $settings['xen_migrate_ip'] . $extrasshsettings . ' ' . '\'' . '(pv -n /dev/' . $dev . '  | gzip -c | dd of="' . $settings['mount_migrate'] . '/xen' . $disknumber . '.img.gz" bs=1M conv=notrunc,noerror status=none) > ' . $progressfilename . '.gz 2>&1 &' . '\'';//trailing & sends to background
+
+					}
 
 					sb_log('Xen - Imaging - /dev/' . $dev);