+++ ChangeLog.md
@@ -3,6 +3,11 @@
  
  [Back To ReadMe](https://github.com/zipurman/oVIRT_Simple_Backup/)
  
+  - 0.6.16 - 2018/04/06
+      - [x] Add a scripted installer for Debian 9
+      - [x] Compression on backup images added as an option
+
+      
   - 0.6.15 - 2018/04/06
       - [x] Added Multi Schedule Manager for Scheduled Backups